slow but steady
